chore: migrate completion to ESM
