--- src/lib/completion/constants.cjs
@@ -1,6 +1,0 @@
-// @ts-check
-const { getPathInHome } = require('../settings.cjs')
-
-const AUTOCOMPLETION_FILE = getPathInHome(['autocompletion.json'])
-
-module.exports = { AUTOCOMPLETION_FILE }
