+++ src/lib/completion/constants.mjs
@@ -1,0 +1,4 @@
+// @ts-check
+import { getPathInHome } from '../settings.cjs'
+
+export const AUTOCOMPLETION_FILE = getPathInHome(['autocompletion.json'])
