--- src/lib/completion/index.cjs
@@ -1,5 +1,0 @@
-const { createAutocompletion } = require('./generate-autocompletion.cjs')
-
-module.exports = {
-  createAutocompletion,
-}
