+++ src/lib/completion/index.mjs
@@ -1,0 +1,1 @@
+export { default as generateAutocompletion } from './generate-autocompletion.mjs'
