+++ tests/unit/lib/completion/snapshots/generate-autocompletion.test.mjs.md
@@ -1,0 +1,74 @@
+# Snapshot report for `tests/unit/lib/completion/generate-autocompletion.test.mjs`
+
+The actual snapshot is saved in `generate-autocompletion.test.mjs.snap`.
+
+Generated by [AVA](https://avajs.dev).
+
+## should generate a completion file
+
+> Snapshot 1
+
+    {
+      bake: {
+        description: 'Cooks something',
+        name: 'bake',
+        options: [
+          {
+            description: 'cook it fast',
+            name: '--fast',
+          },
+          {
+            description: 'Print debugging information',
+            name: '--debug',
+          },
+          {
+            description: 'Proxy server address to route requests through.',
+            name: '--httpProxy',
+          },
+          {
+            description: 'Certificate file to use when connecting using a proxy server',
+            name: '--httpProxyCertificateFilename',
+          },
+        ],
+      },
+      'bake:pizza': {
+        description: 'bakes a pizza',
+        name: 'bake:pizza',
+        options: [
+          {
+            description: 'Type of pizza',
+            name: '--type',
+          },
+          {
+            description: 'Print debugging information',
+            name: '--debug',
+          },
+          {
+            description: 'Proxy server address to route requests through.',
+            name: '--httpProxy',
+          },
+          {
+            description: 'Certificate file to use when connecting using a proxy server',
+            name: '--httpProxyCertificateFilename',
+          },
+        ],
+      },
+      taste: {
+        description: 'tastes something',
+        name: 'taste',
+        options: [
+          {
+            description: 'Print debugging information',
+            name: '--debug',
+          },
+          {
+            description: 'Proxy server address to route requests through.',
+            name: '--httpProxy',
+          },
+          {
+            description: 'Certificate file to use when connecting using a proxy server',
+            name: '--httpProxyCertificateFilename',
+          },
+        ],
+      },
+    }
